+++ src/main/kotlin/com/bugsnag/android/gradle/BugsnagPlugin.kt
@@ -40,6 +40,7 @@ import org.gradle.api.file.RegularFile
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.TaskProvider
 import java.io.File
+import java.net.URL
 import java.util.UUID
 
 /**
@@ -482,7 +483,7 @@ class BugsnagPlugin : Plugin<Project> {
             bundleJsFileProvider.set(File(rnBundle))
             sourceMapFileProvider.set(File(rnSourceMap))
             overwrite.set(bugsnag.overwrite)
-            endpoint.set(bugsnag.endpoint)
+            endpoint.set(getSourcemapUploadEndpoint(bugsnag))
             devEnabled.set("true" == dev)
             failOnUploadError.set(bugsnag.failOnUploadError)
 
@@ -494,6 +495,16 @@ class BugsnagPlugin : Plugin<Project> {
             val cliPath = File(nodeModulesDir, "@bugsnag/source-maps/bin/cli")
             bugsnagSourceMaps.set(cliPath)
             mustRunAfter(rnTask)
+        }
+    }
+
+    internal fun getSourcemapUploadEndpoint(bugsnag: BugsnagPluginExtension): String? {
+        return when (val endpoint = bugsnag.endpoint.get()) {
+            UPLOAD_ENDPOINT_DEFAULT -> endpoint
+            else -> {
+                val host = URL(endpoint)
+                return URL(host, "react-native-source-map").toString()
+            }
         }
     }